fix gating & UI
