+++ packages/core-mobile/app/new/features/swap/hooks/useQuickSwaps.ts
@@ -26,7 +26,16 @@ const SOFTWARE_WALLET_TYPES: ReadonlySet<WalletType> = new Set([
 ])
 
 type UseQuickSwapsResult = {
+  // PostHog feature flag is on; controls whether the Settings section
+  // is rendered at all.
+  flagOn: boolean
+  // Eligibility split into wallet + chain so the Settings UI can show
+  // a specific reason when the toggle is disabled.
+  walletAllowed: boolean
+  chainAllowed: boolean
+  // flagOn && walletAllowed && chainAllowed — the toggle is interactive.
   isAvailable: boolean
+  // isAvailable && saved-settings toggle is on — the bypass actually fires.
   isEnabled: boolean
   feeSetting: QuickSwapFeeLevel
   maxBuy: QuickSwapMaxBuy
@@ -54,6 +63,9 @@ export const useQuickSwaps = (): UseQuickSwapsResult => {
   )
 
   return {
+    flagOn,
+    walletAllowed,
+    chainAllowed,
     isAvailable,
     isEnabled: isAvailable && rawIsEnabled,
     feeSetting,
